requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.20.0
pillow==10.4.0
numpy==1.26.4
gunicorn==22.0.0
h5py==3.11.0
pymongo[srv]==4.9.1
flask-jwt-extended==4.6.0
pycryptodome==3.20.0
python-dotenv==1.0.1
scikit-learn==1.5.2